+++ clang/lib/Sema/SemaType.cpp
@@ -6569,18 +6569,18 @@ static void HandleAddressSpaceTypeAttribute(QualType &Type,
       Attr.setInvalid();
   } else {
     // The keyword-based type attributes imply which address space to use.
-    ASIdx = S.getLangOpts().SYCLIsDevice ? Attr.asSYCLLangAS()
-                                         : Attr.asOpenCLLangAS();
+    ASIdx =
+        S.getLangOpts().isSYCL() ? Attr.asSYCLLangAS() : Attr.asOpenCLLangAS();
     if (S.getLangOpts().HLSL)
       ASIdx = Attr.asHLSLLangAS();
 
     if (ASIdx == LangAS::Default) {
-      assert((S.getLangOpts().SYCLIsDevice || S.getLangOpts().HLSL) &&
+      assert((S.getLangOpts().isSYCL() || S.getLangOpts().HLSL) &&
              "Unexpected language mode");
       S.Diag(Attr.getLoc(), diag::err_attribute_not_supported_in_lang)
           << Attr
-          << (S.getLangOpts().SYCLIsDevice ? AttributeLangSupport::SYCLDevice
-                                           : AttributeLangSupport::HLSL);
+          << (S.getLangOpts().isSYCL() ? AttributeLangSupport::SYCL
+                                       : AttributeLangSupport::HLSL);
       Attr.setInvalid();
       return;
     }